+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 * demisto-sdk lint - Copy dir better handling.
 * demisto-sdk lint - Add error when package missing in docker image.
+* Added *-a , --validate-all* option in *validate* to run all validation on all files.
+
 
 #### 0.5.2
 * Added *-c, --command* option in *generate-docs* to generate a specific command from an integration.